+++ Kata.Checkout/ICheckoutService.cs
@@ -4,5 +4,5 @@ public interface ICheckoutService
 {
     void ScanItem(string item);
     void RemoveItem(string item);
-    float Total();
+    int Total();
 }
