+++ Kata.Checkout/ProductCatalogue/Item.cs
@@ -1,10 +1,10 @@
 ﻿namespace Kata.Checkout.Items;
 
-public class Item(float unitPrice) : IItem
+public class Item(int unitPrice) : IItem
 {
     public string Name { get; }
 
-    public virtual float CalculatePrice(int quantity)
+    public virtual int CalculatePrice(int quantity)
     {
         return unitPrice * quantity;
     }
